# examples/yaak/atr-group-read.openapi.yaml
openapi: 3.1.0
info:
  title: ATR Producer Bundle 1 Demo
  version: 0.1.0
  description: |
    Server-only ATR-shaped producer demo for Group read and read access to
    resources referenced by Group.member.
    Import this file into Yaak (Import -> OpenAPI) to generate requests.
servers:
  - url: http://localhost:{port}/fhir
    description: Local dev server
    variables:
      port:
        default: '3301'
        description: API port (use the port shown in `npm run dev` output if random)
tags:
  - name: Group
  - name: Metadata
  - name: References
paths:
  /Group/{id}:
    get:
      tags: [Group]
      summary: Read ATR Group by id
      description: Returns full ATR-shaped Group. Use `_summary=true` to omit `member`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-group-001
        - name: _summary
          in: query
          required: false
          schema:
            type: string
            enum: ['true']
      responses:
        '200':
          description: Group response
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Missing or invalid bearer token
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Group not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /Patient/{id}:
    get:
      tags: [References]
      summary: Read referenced Patient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-patient-001
      responses:
        '200':
          description: Patient
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /Coverage/{id}:
    get:
      tags: [References]
      summary: Read referenced Coverage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-coverage-001
      responses:
        '200':
          description: Coverage
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Coverage'
        '401':
          description: Unauthorized
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /PractitionerRole/{id}:
    get:
      tags: [References]
      summary: Read referenced PractitionerRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-pcp-001
      responses:
        '200':
          description: PractitionerRole
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/PractitionerRole'
        '401':
          description: Unauthorized
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /Practitioner/{id}:
    get:
      tags: [References]
      summary: Read referenced Practitioner
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-pcp-002
      responses:
        '200':
          description: Practitioner
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Practitioner'
        '401':
          description: Unauthorized
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /Organization/{id}:
    get:
      tags: [References]
      summary: Read referenced Organization
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: demo-provider-org-001
      responses:
        '200':
          description: Organization
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Not found
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'

  /metadata:
    get:
      tags: [Metadata]
      summary: Read minimal FHIR metadata
      responses:
        '200':
          description: CapabilityStatement
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/CapabilityStatement'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    Group:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: Group
        id:
          type: string
          example: demo-group-001
    Patient:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: Patient
        id:
          type: string
          example: demo-patient-001
    Coverage:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: Coverage
        id:
          type: string
          example: demo-coverage-001
    PractitionerRole:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: PractitionerRole
        id:
          type: string
          example: demo-pcp-001
    Practitioner:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: Practitioner
        id:
          type: string
          example: demo-pcp-002
    Organization:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: Organization
        id:
          type: string
          example: demo-payer
    OperationOutcome:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: OperationOutcome
    CapabilityStatement:
      type: object
      additionalProperties: true
      properties:
        resourceType:
          type: string
          example: CapabilityStatement